Описание схемы БД.

+++ internal/delivery/ws/chatting.go
@@ -21,7 +21,7 @@ func (ws *WSHandler) chatting(c *gin.Context) {
 	}
 	defer conn.Close()
 
-	chats, err := ws.Chatting.GetChats(uint64(userID))
+	chats, err := ws.Chatting.GetChats(uint(userID))
 	if err != nil {
 		response.NewError(c, http.StatusInternalServerError, err.Error())
 		return
@@ -40,9 +40,9 @@ func (ws *WSHandler) chatting(c *gin.Context) {
 }
 
 type messageRequest struct {
-	ID     uint64 `json:"message_id"`
-	ChatID uint64 `json:"chat_id"`
-	UserID uint64 `json:"user_id"`
+	ID     uint   `json:"message_id"`
+	ChatID uint   `json:"chat_id"`
+	UserID uint   `json:"user_id"`
 	Text   string `json:"text"`
 }
