+++ internal/models/chat.go
@@ -1,6 +1,6 @@
 package models
 
 type Chat struct {
-	ID       uint64 `json:"chat_id"`
+	ID       uint   `json:"chat_id"`
 	ChatName string `json:"chat_name"`
 }
