+++ internal/models/message.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type Message struct {
-	ID       uint64    `json:"message_id"`
-	ChatID   uint64    `json:"chat_id"`
-	UserID   uint64    `json:"user_id"`
+	ID       uint      `json:"message_id"`
+	ChatID   uint      `json:"chat_id"`
+	UserID   uint      `json:"user_id"`
 	Text     string    `json:"text"`
 	SendTime time.Time `json:"send_time"`
 }
