+++ internal/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID       uint64 `json:"-"`
+	ID       uint   `json:"-"`
 	Name     string `json:"name"`
 	UserName string `json:"username"`
 	Email    string `json:"email"`
