+++ internal/service/service.go
@@ -13,7 +13,7 @@ type Chat interface {
 }
 
 type Chatting interface {
-	GetChats(userID uint64) ([]models.Chat, error)
+	GetChats(userID uint) ([]models.Chat, error)
 	GetHistory(string) ([]string, error)
 	Publish(message *models.Message) error
 	Subscribe(chatIDs []string) *redis.PubSub
